# back/openapi.yaml
...
openapi: 3.0.3
info:
  title: 파이어가드 API
  version: 0.7.0
  description: |
    CCTV 화재 감지 → 관리자 알림 → 무응답 시 119 자동 신고·승계 백엔드 API.

    ## 공통 규칙
    - 모든 필드는 `snake_case`, 날짜·시각은 ISO 8601 (`2026-08-10T14:30:00`)
    - 인증은 `Authorization: Bearer <access_token>` (로그인·헬스체크·미디어 제외)
    - AI 모델 전용 `/api/internal/*` 은 JWT 대신 `X-Internal-Key` 헤더를 쓴다
    - 에러는 어떤 상황에서도 `{ "code": "...", "message": "..." }` 형태다.
      검증 실패(400)는 어느 입력 탓인지 가리킬 수 있으면 `field` 가 하나 더 붙는다
      (`{ "code": "BAD_REQUEST", "message": "...", "field": "user_pw" }`) —
      400 은 `code` 가 전부 `BAD_REQUEST` 라, 폼 입력칸에 메시지를 붙일 때 이걸 쓴다
    - 목록 중 페이징이 있는 것은 `items / page / size / total_count / total_pages`
      봉투로 감싸 내려간다 (`/api/cctvs`, `/api/agencies` 는 페이징 없이 `items` 만)

    ## 실시간 알림 — 폴링하지 않는다 (2026-08-11 확정)
    WebSocket·SSE 뿐 아니라 **폴링도 쓰지 않는다**. 119 신고가 화재확인(`POST /api/alerts/{alert_no}/respond`)
    요청 안에서 **동기로** 끝나므로, 기다렸다가 알려줄 상태 변화가 없다.
    - 사람을 깨우는 채널은 **SMS** 다. 문자를 보고 앱을 열면 그때 목록을 한 번 조회하면 된다.
    - 남은 시간 카운트다운은 응답에 실린 `alert_deadline_at` 으로 클라이언트가 로컬 계산한다.
    - 화면 갱신 원칙: **열 때 조회 + 행동 후 갱신**.

    ## 판정·유예 기본값 (2026-08-10 변경)
    | 값 | 기본 | 뜻 |
    |---|---|---|
    | `EVENT_WINDOW_SEC` | 60초 | 관측 창. **최초 감지 시각에 고정**되며 이후 검출로 연장되지 않는다 |
    | `EVENT_THRESHOLD_FRAMES` | 30 | 창 안에서 이만큼 쌓이면 화재 확정. 미달인 채 창이 닫히면 `DISMISSED` |
    | `ALERT_DEADLINE_SEC` | 30초 | 알림 응답 유예. **이전 `ALERT_DEADLINE_MIN`(3분)에서 바뀐 값** |
    | `ESCALATION_INTERVAL_SEC` | 5초 | 무응답 스윕 주기. 신고가 최대 이만큼 늦어져 실제 유예는 30~35초 |

...
servers:
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: 로그인
      description: |
        아이디·비밀번호로 JWT 를 발급받는다. 정지(SUSPENDED)·탈퇴(WITHDRAWN)
        계정은 비밀번호가 맞아도 403 이다.
        로그인에는 비밀번호 작성규칙을 적용하지 않는다 (기존 계정 보장).

        소셜 로그인으로 가입한 계정은 비밀번호가 없어서 이 경로로는 절대 들어올 수
        없다 — 400 `SOCIAL_ACCOUNT` 로 돌려보낸다. 어느 서비스로 가입했는지까지는
        알려주지 않으니, 프론트는 소셜 로그인 버튼들을 다시 보여주면 된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, user_pw]
              properties:
                user_id:
                  type: string
                  example: admin01
                user_pw:
                  type: string
                  format: password
                  example: "Fire#123"
      responses:
        "200":
          description: 발급 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: Bearer 토큰
                  user:
                    $ref: "#/components/schemas/LoginUser"
        "400":
          description: 필수값 누락·형식 오류(`BAD_REQUEST`) 또는 소셜 계정(`SOCIAL_ACCOUNT`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missing_field:
                  summary: 필수값 누락
                  value:
                    code: BAD_REQUEST
                    message: user_id 는 필수입니다.
                    field: user_id
                social:
                  summary: 소셜 로그인 계정
                  value:
                    code: SOCIAL_ACCOUNT
                    message: 소셜 로그인으로 가입한 계정입니다. 해당 서비스로 로그인해주세요.
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/AccountBlocked"
  /api/auth/logout:
    post:
      tags: [auth]
      summary: 로그아웃 (토큰 폐기 + 이력 기록)
      description: |
        **토큰을 폐기하고** 활동이력에 `LOGOUT` 을 1행 남긴다. 두 가지를 다 한다.

        폐기는 사용자 단위다 — **그 사람의 다른 기기도 함께 끊긴다.** 이후 기존
        토큰으로 보낸 요청은 401 `TOKEN_REVOKED` 를 받는다. 로그아웃 직후 바로
        다시 로그인하는 것은 정상 동작한다.

        이 엔드포인트를 불러야만 둘 다 일어난다. 프론트가 저장한 토큰만 지우고
        이걸 건너뛰면 서버 쪽 토큰은 만료(기본 12시간)까지 살아있고 `LOGOUT` 행도
        쌓이지 않는다. 브라우저를 그냥 닫은 경우도 마찬가지로 잡히지 않는다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 폐기·기록 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: 로그아웃되었습니다.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/me:
    get:
      tags: [auth]
      summary: 내 정보
      description: 새로고침 후 세션 복원용. 토큰의 주인 정보를 DB 에서 다시 읽어 준다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/UserNotFound"
  /api/auth/find-id:
    post:
      tags: [auth]
      summary: 아이디 찾기
      description: |
        이름 + 이메일이 일치하면 아이디를 **부분 마스킹**해 돌려준다 (`admin01` → `adm****`).

        통째로 주지 않는 이유: 이름과 이메일은 팀 안에서 대개 알려진 정보라,
        전체를 돌려주면 그 조합을 아는 사람에게 계정 목록을 넘기는 셈이 된다.
        본인이라면 앞 3자로 알아본다.

        **탈퇴·정지 계정은 찾아지지 않는다** (404). 되찾아도 로그인이 막혀 있고,
        비활성 계정의 존재를 확인해 주는 것 자체가 불필요한 노출이다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_name, user_email]
              properties:
                user_name:
                  type: string
                  example: 관리자
                user_email:
                  type: string
                  format: email
                  example: admin@fg.kr
      responses:
        "200":
          description: 일치하는 계정을 찾음
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    description: 앞 3자만 남기고 마스킹된 아이디
                    example: adm****
        "400":
          description: 필수값 누락 (`BAD_REQUEST`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: BAD_REQUEST
                message: user_email 는 필수입니다.
                field: user_email
        "404":
          description: 일치하는 계정 없음 (`USER_NOT_FOUND`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: USER_NOT_FOUND
                message: 일치하는 계정을 찾을 수 없습니다.
  /api/auth/password-reset/request:
    post:
      tags: [auth]
      summary: 비밀번호 재설정 — 인증코드 발급
      description: |
        아이디 + 이름 + 이메일이 일치하면 인증코드를 만들어 **등록된 연락처로 SMS 발송**한다.
        받은 코드로 `/api/auth/password-reset/confirm` 을 부르면 비밀번호가 바뀐다.

        ⚠️ **정보가 틀려도 200 이 나간다.** 404 를 내면 조합을 바꿔가며 찔러 실재 계정을
        골라낼 수 있어서다. 응답만으로는 계정 존재 여부를 알 수 없다 — 프론트는
        "정보가 일치하면 문자를 보냈다"는 안내만 띄우고 다음 단계로 넘기면 된다.

        ⚠️ **코드는 응답 본문에 실리지 않는다.** 넣으면 인증 단계가 무의미해진다.
        개발 중에는 목(mock) SMS 발송기가 서버 로그에 찍으므로 그것으로 확인하면 된다.

        코드는 서버에 저장되지 않고 매 요청마다 계산해 대조한다. 그래서
        **비밀번호가 바뀌는 순간 이전 코드는 자동으로 무효**가 된다 (1회용).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, user_name, user_email]
              properties:
                user_id:
                  type: string
                  example: admin01
                user_name:
                  type: string
                  example: 관리자
                user_email:
                  type: string
                  format: email
                  example: admin@fg.kr
      responses:
        "200":
          description: 접수됨 (정보 일치 여부와 무관하게 같은 응답이다)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 정보가 일치하면 등록된 연락처로 인증코드를 보냈습니다.
                  expires_in_sec:
                    type: integer
                    description: 코드 유효시간(초). 발급 시점에 따라 최대 30초 짧아질 수 있다
                    example: 300
        "400":
          description: 필수값 누락 (`BAD_REQUEST`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: BAD_REQUEST
                message: user_email 는 필수입니다.
                field: user_email
  /api/auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: 비밀번호 재설정 — 인증코드 확인 후 변경
      description: |
        인증코드가 맞으면 새 비밀번호로 바꾼다. 성공하면 새 비밀번호로 바로 로그인하면 된다.

        - 코드는 **대소문자·공백을 가리지 않는다**. 문자로 받은 값을 소문자로 치거나
          중간에 공백을 넣어도 통과한다 (`ABCDE FGHJK` = `abcdefghjk`).
        - 비밀번호 작성규칙은 **가입할 때와 똑같이** 적용된다. 재설정으로 규칙을
          우회할 수 있으면 규칙을 둔 의미가 없다.
        - 코드가 틀렸을 때와 아이디가 없을 때의 응답이 **같다** (`INVALID_RESET_CODE`).
          아이디 존재 여부를 흘리지 않기 위해서다.
        - 성공하면 방금 쓴 코드는 그 자리에서 무효가 된다 (해시가 바뀌므로 재사용 불가).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, code, user_pw]
              properties:
                user_id:
                  type: string
                  example: admin01
                code:
                  type: string
                  description: SMS 로 받은 10자리 인증코드
                  example: KQ7MB2XZP4
                user_pw:
                  type: string
                  format: password
                  description: 새 비밀번호. 영문+숫자+특수문자 3종 포함 8~64자
                  example: "Reset#2026x"
      responses:
        "200":
          description: 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 비밀번호가 변경되었습니다. 새 비밀번호로 로그인해주세요.
        "400":
          description: 인증코드 불일치·만료 또는 비밀번호 규칙 위반
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalid_code:
                  summary: 코드가 틀렸거나 만료됨 (아이디가 없을 때도 같은 응답)
                  value:
                    code: INVALID_RESET_CODE
                    message: 인증코드가 올바르지 않거나 만료되었습니다.
                password_rule:
                  summary: 새 비밀번호가 작성규칙 위반
                  value:
                    code: BAD_REQUEST
                    message: 비밀번호는 영문자·숫자·특수문자를 모두 포함해 8자 이상이어야 합니다.
                    field: user_pw
  /api/auth/oauth/{provider}/authorize:
    get:
      tags: [auth]
      summary: 소셜 로그인 — 동의화면 주소 발급
      description: |
        구글·카카오·네이버 동의화면 주소와 `state` 를 만들어 준다. 로그인 화면에서
        부르는 **비로그인 공개 엔드포인트**다.

        프론트 순서:
        1. 이 API 를 부른다
        2. `state` 를 `sessionStorage` 에 넣는다
        3. `authorize_url` 로 이동한다
        4. 콜백 화면(`/oauth/callback/{provider}`)에서 받은 `code` 와 저장해 둔
           `state` 를 `POST /api/auth/oauth/{provider}` 에 보낸다

        `redirect_uri` 는 **서버가 계산해서 authorize_url 에 넣는다** —
        `{OAUTH_REDIRECT_BASE}/oauth/callback/{provider}`. 응답에 따로 싣지 않는 것은
        프론트가 그 값을 자기 쪽에도 적어 두면 정의가 두 곳이 되기 때문이다.
        각 개발자 콘솔에 **똑같은 URI 를 등록해야** 한다 (안 하면 프로바이더가
        `redirect_uri_mismatch` 로 돌려보낸다).

        `state` 는 서버에 저장되지 않는다. HMAC 으로 계산해 대조하며 약 10분 뒤
        만료된다 — 동의화면을 열어 둔 채 오래 있으면 400 `INVALID_OAUTH_STATE` 다.
      parameters:
        - $ref: "#/components/parameters/OauthProvider"
      responses:
        "200":
          description: 발급 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorize_url:
                    type: string
                    description: 사용자를 보낼 프로바이더 동의화면 주소 (state 가 실려 있다)
                    example: https://accounts.google.com/o/oauth2/v2/auth?response_type=code&client_id=...&state=...
                  state:
                    type: string
                    description: sessionStorage 에 보관했다가 콜백에서 그대로 돌려줄 값
                    example: 3Qk1t7Yz9Ab2.9f86d081884c7d659a2feaa0c55ad015a3bf4f1b
        "400":
          $ref: "#/components/responses/UnsupportedProvider"
        "503":
          $ref: "#/components/responses/OauthNotConfigured"
  /api/auth/oauth/{provider}:
    post:
      tags: [auth]
      summary: 소셜 로그인 — code 교환 후 로그인
      description: |
        콜백에서 받은 `code` 를 프로바이더에서 토큰·사용자정보로 바꾸고, 우리 JWT 를
        발급한다. 응답은 `POST /api/auth/login` 과 **완전히 같은 모양**이라 프론트는
        두 버튼의 결과를 같은 코드로 받으면 된다.

        처음 들어온 계정은 그 자리에서 만들어진다 —
        `user_role: VIEWER`, `user_status: ACTIVE`, 비밀번호 없음.
        관제 권한(`ADMIN`)이 필요하면 관리자가 따로 올려준다.

        ⚠️ **이메일이 같은 일반 계정과 자동으로 연결되지 않는다.** 프로바이더가 그
        이메일의 소유를 실제로 검증했는지 보장할 수 없어서다. 검증하지 않는 곳에
        피해자의 이메일로 가입해 로그인하면 우리 쪽 계정을 그대로 가져가게 된다.
        그래서 `admin@fg.kr` 로 쓰는 계정이 이미 있어도 **별개 계정**이 생긴다.

        계정 두 개를 합치는 것은 본인 확인을 거친 별도 기능이어야 할 문제라
        이 API 의 범위 밖이다.

        비밀번호가 없는 계정이라 `PUT /api/users/password` 와
        `/api/auth/password-reset/*` 는 쓸 수 없다 (400 `SOCIAL_ACCOUNT`).
        로그아웃(`POST /api/auth/logout`)은 일반 계정과 똑같이 동작한다.
      parameters:
        - $ref: "#/components/parameters/OauthProvider"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, state]
              properties:
                code:
                  type: string
                  description: 프로바이더가 콜백 쿼리스트링으로 준 authorization code
                  example: 4/0AeanS0b...
                state:
                  type: string
                  description: authorize 응답으로 받아 sessionStorage 에 넣어 둔 값
                  example: 3Qk1t7Yz9Ab2.9f86d081884c7d659a2feaa0c55ad015a3bf4f1b
      responses:
        "200":
          description: 로그인 성공 (신규 계정이면 이때 만들어진다)
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: Bearer 토큰
                  user:
                    $ref: "#/components/schemas/LoginUser"
        "400":
          description: |
            지원하지 않는 프로바이더(`UNSUPPORTED_PROVIDER`) · 필수값 누락
            (`BAD_REQUEST`) · state 위조·만료(`INVALID_OAUTH_STATE`).

            `INVALID_OAUTH_STATE` 는 사용자가 잘못한 게 아닌 경우가 많다 —
            동의화면을 오래 열어 뒀거나 다른 탭에서 이어받은 경우다.
            로그인 버튼부터 다시 누르게 안내하면 된다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unsupported:
                  summary: 지원하지 않는 프로바이더
                  value:
                    code: UNSUPPORTED_PROVIDER
                    message: 지원하지 않는 소셜 로그인입니다.
                missing_field:
                  summary: 필수값 누락
                  value:
                    code: BAD_REQUEST
                    message: code 는 필수입니다.
                    field: code
                invalid_state:
                  summary: state 위조·만료
                  value:
                    code: INVALID_OAUTH_STATE
                    message: 인증 요청이 만료되었거나 올바르지 않습니다. 다시 시도해주세요.
        "403":
          $ref: "#/components/responses/AccountBlocked"
        "409":
          description: |
            합성한 `user_id`(`{provider}_{provider_id}` 를 50자로 자른 값)가 다른
            계정과 겹쳤다 (`DUPLICATE_USER_ID`). 자동으로 이름을 바꿔 만들지 않는다 —
            계정이 조용히 늘어나는 것보다 사람이 보는 편이 낫다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: DUPLICATE_USER_ID
                message: 이미 사용 중인 아이디입니다.
        "502":
          description: |
            프로바이더와의 통신 실패 (`OAUTH_PROVIDER_ERROR`). 네트워크 오류,
            만료·재사용된 `code`, 프로바이더 장애가 전부 여기로 들어온다.
            네이버는 실패해도 HTTP 200 을 주기 때문에 본문 `resultcode` 로 판정한 것도 포함한다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: OAUTH_PROVIDER_ERROR
                message: 소셜 로그인 제공자와 통신하지 못했습니다. 잠시 후 다시 시도해주세요.
        "503":
          $ref: "#/components/responses/OauthNotConfigured"

  # ---------------------------------------------------------------- 계정
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '로그인으로 받은 `access_token` 을 `Authorization: Bearer <token>` 으로 보낸다.'
    internalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: AI 모델 전용. 서버의 `INTERNAL_API_KEY` 환경변수 값과 대조한다.
  parameters:
...
    OauthProvider:
      name: provider
      in: path
      required: true
      description: |
        소셜 로그인 제공자. 경로에는 **소문자**로 쓴다 (대문자로 보내도 받아준다).
        DB 의 `user_provider` 에는 항상 대문자로 저장된다.
      schema:
        type: string
        enum: [google, kakao, naver]
        example: google

  # 에러 응답은 반드시 example 을 단다 —
  # 없으면 Swagger UI 가 공통 Error 스키마로 본문을 만들어 내서 400·401·403 이
  # 전부 똑같이 보이고, 프론트가 분기할 code 를 화면에서 알아낼 수 없다.
  # (tests/test_openapi.py::test_error_responses_declare_example 가 강제한다)
  # 400 은 공통 응답을 두지 않는다 — 엔드포인트마다 실제로 나가는 문구와 field 가
  # 다르고, 공통 예시 하나를 돌려쓰면 카메라 등록 400 에 "user_id 는 필수입니다"
  # 같은 거짓 예시가 붙는다. 400 은 오퍼레이션 안에 직접 쓴다.
  responses:
    Unauthorized:
      description: |
        토큰 없음(`UNAUTHORIZED`) · 만료(`TOKEN_EXPIRED`) · 손상(`INVALID_TOKEN`) ·
        폐기(`TOKEN_REVOKED`). 넷 다 로그인 화면으로 돌려보내야 하지만,
        **안내 문구는 갈라주는 편이 낫다.**

        `TOKEN_REVOKED` 는 사용자가 잘못한 게 없는데 튕기는 경우다 — 다른 기기에서
        로그아웃했거나 관리자가 계정을 정지했을 때 나온다. 이유를 안 알려주면
        사용자는 영문도 모르고 튕긴 것으로 느낀다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missing:
              summary: 토큰 없음
              value:
                code: UNAUTHORIZED
                message: 인증 토큰이 필요합니다.
            expired:
              summary: 토큰 만료
              value:
                code: TOKEN_EXPIRED
                message: 토큰이 만료되었습니다. 다시 로그인해주세요.
            invalid:
              summary: 토큰 손상
              value:
                code: INVALID_TOKEN
                message: 유효하지 않은 토큰입니다.
            revoked:
              summary: 토큰 폐기 (로그아웃·계정 정지)
              value:
                code: TOKEN_REVOKED
                message: 로그아웃되었거나 무효화된 토큰입니다. 다시 로그인해주세요.
...
    InvalidCredentials:
      description: |
        아이디 또는 비밀번호 불일치 (`INVALID_CREDENTIALS`).
        존재하지 않는 아이디와 비밀번호 오류를 **구분하지 않는다** —
        구분하면 계정 존재 여부를 확인하는 통로가 된다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INVALID_CREDENTIALS
            message: 아이디 또는 비밀번호가 일치하지 않습니다.
    UnsupportedProvider:
      description: 지원하지 않는 소셜 로그인 (`UNSUPPORTED_PROVIDER`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: UNSUPPORTED_PROVIDER
            message: 지원하지 않는 소셜 로그인입니다.
    OauthNotConfigured:
      description: |
        서버에 그 프로바이더의 앱 키(`*_CLIENT_ID` / `*_CLIENT_SECRET`)가 없다
        (`OAUTH_NOT_CONFIGURED`). 키를 안 넣고 배포해도 서버는 그대로 뜨고 나머지
        로그인 경로도 살아 있다 — 그 버튼만 막힌다.

        500 이 아니라 503 으로 따로 내는 이유: 500 이면 로그만 보고 원인을 찾을 수
        없고, 프론트도 "아직 준비되지 않음"으로 안내할 방법이 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: OAUTH_NOT_CONFIGURED
            message: 해당 소셜 로그인이 서버에 설정되어 있지 않습니다.
    AccountBlocked:
      description: |
        정지·탈퇴 계정. 비밀번호가 맞아야 여기까지 온다 (틀리면 401 이 먼저 나간다).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            suspended:
              summary: 정지된 계정
              value:
                code: ACCOUNT_SUSPENDED
                message: 정지된 계정입니다.
            withdrawn:
              summary: 탈퇴한 계정
              value:
                code: ACCOUNT_WITHDRAWN
                message: 탈퇴한 계정입니다.
...
    UserNotFound:
      description: 사용자 없음 (`USER_NOT_FOUND`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: USER_NOT_FOUND
            message: 사용자를 찾을 수 없습니다.
...
  schemas:
...
    Error:
      type: object
      description: |
        모든 에러 응답의 공통 형식.

        `field` 는 검증 실패(400)에서 **어느 입력 때문인지** 가리킬 수 있을 때만
        실린다. 400 은 `code` 가 전부 `BAD_REQUEST` 하나라, 이게 없으면 프론트가
        폼의 어느 칸에 메시지를 붙일지 정하려고 한글 문구를 파싱해야 한다.
        입력 하나를 특정할 수 없는 400(예: "수정할 필드가 없습니다")과
        400 이 아닌 에러에는 키 자체가 없다.
      required: [code, message]
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          description: 사용자에게 그대로 보여줄 수 있는 한글 메시지
          example: user_id 는 필수입니다.
        field:
          type: string
          description: 문제가 된 요청 필드·쿼리 파라미터 이름 (없을 수 있음)
          example: user_id
...
    UserRole:
      type: string
      enum: [ADMIN, VIEWER]
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, WITHDRAWN]
...
    LoginUser:
      type: object
      description: 로그인 직후 화면에 바로 필요한 최소 정보
      properties:
        user_no:
          type: integer
          format: int64
        user_id:
          type: string
        user_name:
          type: string
          nullable: true
        user_role:
          $ref: "#/components/schemas/UserRole"
    Me:
      type: object
      properties:
        user_no:
          type: integer
          format: int64
        user_id:
          type: string
        user_name:
          type: string
          nullable: true
        user_email:
          type: string
          nullable: true
        user_phone:
          type: string
          nullable: true
          description: 하이픈 없이 숫자만
        user_role:
          $ref: "#/components/schemas/UserRole"
        user_status:
          $ref: "#/components/schemas/UserStatus"
...